pyproject.toml
[build-system]
requires = ["setuptools>=75"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["apps*", "packages*"]

[project]
name = "finalmaxbot"
version = "0.1.0"
description = "MAX messenger AI bot — backend (webhook, API)"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
  "httpx>=0.27.0",
  "pydantic-settings>=2.6.0",
  "sqlalchemy[asyncio]>=2.0.36",
  "aiosqlite>=0.20.0",
  "asyncpg>=0.30.0",
  "alembic>=1.14.0",
  "pillow>=11.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-asyncio>=0.24.0",
  "ruff>=0.8.0",
]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
